V1.0.0
Sure! Here is the repository that contains a LaTeX resume template customized for personal use along with an example of my resume. If you would like to use it to create your own CV, feel free to clone the repository and modify it according to your needs. Feel free to reach out if you need any help!